Moved sidebar components into its own folder.
